+++ scripts/wsgi.py
@@ -1,6 +1,6 @@
 import sys
 sys.path.append("..")
-from service import app
+from src import app
 
 if __name__ == "__main__":
     app.run(debug=True)
