--- service/__init__.py
@@ -1,1 +1,0 @@
-from .fastAPI_project import *
