+++ src.egg-info/PKG-INFO
@@ -1,0 +1,4 @@
+Metadata-Version: 2.1
+Name: src
+Version: 0.0.0
+License-File: LICENSE
