+++ src.egg-info/SOURCES.txt
@@ -1,0 +1,13 @@
+LICENSE
+README.md
+setup.py
+src/__init__.py
+src.egg-info/PKG-INFO
+src.egg-info/SOURCES.txt
+src.egg-info/dependency_links.txt
+src.egg-info/top_level.txt
+src/services/__init__.py
+src/services/service.py
+tests/__init__.py
+tests/results_service.py
+tests/test_service.py
