+++ src.egg-info/top_level.txt
@@ -1,0 +1,2 @@
+src
+tests
