+++ src/__init__.py
@@ -1,0 +1,1 @@
+from .view.fastAPI_project import *
