+++ tests/results_service.py
@@ -1,0 +1,13 @@
+
+
+
+csr1_1 =  [[ 1.77356731e-02 , -2.04058073e-04 , 1.40530020e-02 , 1.04300003e+01,
+   1.07168760e+01,  1.08766742e-01,  1.63000000e+00,  0.00000000e+00,
+   3.39000015e+01,  6.53999996e+00,  5.42999983e+00,  6.30000019e+00,
+   2.45001971e+06,  2.45248294e+06]]   
+csr1_2 = ['RA', 'Dec', 'SideProb', 'Fpeak', 'Fint', 'rms', 'Major', 'Minor', 'PosAng', 'FitMajor', 'FitMinor', 'FitPosAng', 'StartMJD', 'StopMJD'] 
+csr1_3 = ['rad', 'rad', ' ', 'mJy', 'mJy', 'mJy', 'arcsec', 'arcsec', 'deg', 'arcsec', 'arcsec', 'deg', 'MJD', 'MJD'] 
+
+
+scsr1 = {'RA': {'units': 'deg', 'values': [1.0161792162935126]}, 'Dec': {'units': 'deg', 'values': [-0.011691666358208579]}, 'SideProb': {'units': ' ', 'values': [0.014053001999855042]}, 'Fpeak': {'units': 'mJy', 'values': [10.430000305175781]}, 'Fint': {'units': 'mJy', 'values': [10.716876029968262]}, 'rms': {'units': 'mJy', 'values': [0.10876674205064774]}, 'Major': {'units': 'arcsec', 'values': [1.6299999952316284]}, 'Minor': {'units': 'arcsec', 'values': [0.0]}, 'PosAng': {'units': 'deg', 'values': [33.900001525878906]}, 'FitMajor': {'units': 'arcsec', 'values': [6.539999961853027]}, 'FitMinor': {'units': 'arcsec', 'values': [5.429999828338623]}, 'FitPosAng': {'units': 'deg', 'values': [6.300000190734863]}, 'StartMJD': {'units': 'MJD', 'values': [2450019.713368056]}, 'StopMJD': {'units': 'MJD', 'values': [2452482.9428125005]}}
+scsr2 = {'RA': {'units': 'deg', 'values': [1.016179215453705]}, 'Dec': {'units': 'deg', 'values': [-0.011691666358472456]}, 'SideProb': {'units': ' ', 'values': [0.014053002]}, 'Fpeak': {'units': 'mJy', 'values': [10.4300003]}, 'Fint': {'units': 'mJy', 'values': [10.716876]}, 'rms': {'units': 'mJy', 'values': [0.108766742]}, 'Major': {'units': 'arcsec', 'values': [1.63]}, 'Minor': {'units': 'arcsec', 'values': [0.0]}, 'PosAng': {'units': 'deg', 'values': [33.9000015]}, 'FitMajor': {'units': 'arcsec', 'values': [6.53999996]}, 'FitMinor': {'units': 'arcsec', 'values': [5.42999983]}, 'FitPosAng': {'units': 'deg', 'values': [6.30000019]}, 'StartMJD': {'units': 'MJD', 'values': [2450019.71]}, 'StopMJD': {'units': 'MJD', 'values': [2452482.94]}}
