+++ text.txt
@@ -1,0 +1,1 @@
+[[0.0177356731, -0.000204058073, 0.014053002, 10.4300003, 10.716876, 0.108766742, 1.63, 0.0, 33.9000015, 6.53999996, 5.42999983, 6.30000019, 2450019.71, 2452482.94]] ['RA', 'Dec', 'SideProb', 'Fpeak', 'Fint', 'rms', 'Major', 'Minor', 'PosAng', 'FitMajor', 'FitMinor', 'FitPosAng', 'StartMJD', 'StopMJD'] ['rad', 'rad', ' ', 'mJy', 'mJy', 'mJy', 'arcsec', 'arcsec', 'deg', 'arcsec', 'arcsec', 'deg', 'MJD', 'MJD']
