+++ text2.txt
@@ -1,0 +1,1 @@
+{'RA': {'units': 'deg', 'values': [1.016179215453705]}, 'Dec': {'units': 'deg', 'values': [-0.011691666358472456]}, 'SideProb': {'units': ' ', 'values': [0.014053002]}, 'Fpeak': {'units': 'mJy', 'values': [10.4300003]}, 'Fint': {'units': 'mJy', 'values': [10.716876]}, 'rms': {'units': 'mJy', 'values': [0.108766742]}, 'Major': {'units': 'arcsec', 'values': [1.63]}, 'Minor': {'units': 'arcsec', 'values': [0.0]}, 'PosAng': {'units': 'deg', 'values': [33.9000015]}, 'FitMajor': {'units': 'arcsec', 'values': [6.53999996]}, 'FitMinor': {'units': 'arcsec', 'values': [5.42999983]}, 'FitPosAng': {'units': 'deg', 'values': [6.30000019]}, 'StartMJD': {'units': 'MJD', 'values': [2450019.71]}, 'StopMJD': {'units': 'MJD', 'values': [2452482.94]}}
